+++ test/environments/staticarrays/src/PureRNGsStaticArraysConformance.jl
@@ -1,0 +1,1 @@
+module PureRNGsStaticArraysConformance end